fix: rename DestinationTypeExtended to DestinationType DestinationType to ParticipantType
