+++ src/Sinch/Voice/Common/Destination.cs
@@ -14,9 +14,9 @@ namespace Sinch.Voice.Common
         /// </summary>
         [JsonPropertyName("type")]
 #if NET7_0_OR_GREATER
-        public required DestinationType Type { get; set; }
+        public required ParticipantType Type { get; set; }
 #else 
-        public DestinationType Type { get; set; } = null!;
+        public ParticipantType Type { get; set; } = null!;
 #endif
 
         /// <summary>
